[align]: (3) moved functions into library/funcitons/image_processing, added type annotations and fixed some type errors
